requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
pydantic-settings
httpx
pytest
pydantic
rich
python-json-logger
asgi-correlation-id
python-jose
python-multipart
passlib[bcrypt]
starlette
pytest-mock
b2sdk
aiofiles
openai
mysqlclient
aiomysql